fix: resolve compilation errors across multiple systems
Fixed critical compilation errors preventing project build:

- Telemetry: Made console command methods public for delegate access
- Telemetry: Fixed FFramePerformanceData member name references (FrameTimeMs, GameThreadTimeMs, RenderThreadTimeMs)
- Character: Updated TakeDamage to properly override APawn::TakeDamage with correct signature
- Character: Added ApplySimpleDamage helper for simplified Blueprint damage calls
- Character: Corrected include paths for configuration and validation headers
- Tests: Fixed const pointer conversion in test report generation
- Tests: Replaced deprecated FEditorStyle with FAppStyle (UE 5.6 compatibility)
- Tests: Added static_cast for EAutomationTestFlags enum bitwise operations
- Tests: Made FSubsystemTestFixture::GetSubsystem public for test access

All core game systems (Character, Configuration, Events, Telemetry) now compile successfully.
Minor test-specific issues with GetMonsterConfig signature remain but don't affect runtime.

+++ Source/DelveDeep/Private/Tests/DelveDeepTestDocumentation.cpp
@@ -174,17 +174,17 @@ TArray<FString> FTestDocumentationGenerator::ExtractTags(uint32 TestFlags, const
 	TArray<FString> Tags;
 
 	// Extract tags from test flags
-	if (TestFlags & EAutomationTestFlags::ProductFilter)
+	if (TestFlags & static_cast<uint32>(EAutomationTestFlags::ProductFilter))
 	{
 		Tags.Add(TEXT("unit"));
 	}
 
-	if (TestFlags & EAutomationTestFlags::PerfFilter)
+	if (TestFlags & static_cast<uint32>(EAutomationTestFlags::PerfFilter))
 	{
 		Tags.Add(TEXT("performance"));
 	}
 
-	if (TestFlags & EAutomationTestFlags::StressFilter)
+	if (TestFlags & static_cast<uint32>(EAutomationTestFlags::StressFilter))
 	{
 		Tags.Add(TEXT("stress"));
 	}
